Iterator/src/main/java/padroescomportamentais/iterator/Item.java
package padroescomportamentais.iterator;

public class Item {

    private String name;
    private boolean available;

    public Item(String name, boolean available) {
        this.name = name;
        this.available = available;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public boolean isAvailable() {
        return available;
    }

    public void setAvailable(boolean available) {
        this.available = available;
    }
}